requirement.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
django-psycopg2-extension==0.0.4
djangorestframework==3.11.1
environ==1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
